[ADD FONCTIONS UNCODED]

+++ srcs/ft_strcpy.c
@@ -6,15 +6,17 @@
 /*   By: jobailla <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/08/09 11:29:02 by jobailla          #+#    #+#             */
-/*   Updated: 2016/10/05 22:50:02 by jobailla         ###   ########.fr       */
+/*   Updated: 2016/10/12 11:34:01 by jobailla         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+/* Copier une chaîne. */
 
 #include "libft.h"
 
 char	*ft_strcpy(char *dst, const char *src)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (src[i] != '\0')
